added unit tests for the migration and fixed rounding errors
